+++ src/importDataInfo/WorkMoveInfo.java
@@ -4,7 +4,7 @@ package importDataInfo;
  * Created by leko on 2016/1/18.
  */
 public class WorkMoveInfo {
-    private Integer CWPWORKMOVENUM;     //Move的顺序
+    private Long CWPWORKMOVENUM;     //Move的顺序
     private String DECK;                //D甲板下H甲板上
     private Integer GLOBALPRIORITY;     //全为2
     private String HATCH;               //舱
@@ -12,11 +12,11 @@ public class WorkMoveInfo {
     private String LD;                  //D卸船L装船
     private String MOVETYPE;            //作业工艺（暂时写尺寸）
 
-    public Integer getCWPWORKMOVENUM() {
+    public Long getCWPWORKMOVENUM() {
         return CWPWORKMOVENUM;
     }
 
-    public void setCWPWORKMOVENUM(Integer CWPWORKMOVENUM) {
+    public void setCWPWORKMOVENUM(Long CWPWORKMOVENUM) {
         this.CWPWORKMOVENUM = CWPWORKMOVENUM;
     }
